+++ cmplibrary/src/main/java/com/sourcepoint/gdpr_cmplibrary/ConsentLibException.java
@@ -14,7 +14,7 @@ public class ConsentLibException extends Exception {
     }
 
     public static class NoInternetConnectionException extends ConsentLibException {}
-    
+
     public static class ApiException extends ConsentLibException {
         ApiException(Exception e, String message) { super(e, "Error due to android API: " + message); }
         ApiException(String message) { super("Error due to android API: " + message); }