+++ cmplibrary/src/main/java/com/sourcepoint/gdpr_cmplibrary/SourcePointClient.java
@@ -24,7 +24,7 @@ class SourcePointClient {
     private static AsyncHttpClient http = new AsyncHttpClient();
 
     private static final String baseMsgUrl = "https://wrapper-api.sp-prod.net/tcfv2/v1/gdpr/message-url?inApp=true";
-    private static final String baseNativeMsgUrl = "https://fake-wrapper-api.herokuapp.com/tcfv2/v1/gdpr/native-message";
+    private static final String baseNativeMsgUrl = "https://wrapper-api.sp-prod.net/tcfv2/v1/gdpr/native-message?inApp=true";
     private static final String baseSendConsentUrl = "https://wrapper-api.sp-prod.net/tcfv2/v1/gdpr/consent?inApp=true";
 
     private int accountId;
@@ -62,8 +62,9 @@ class SourcePointClient {
         }
     }
 
-    class MessageResponseHandler extends JsonHttpResponseHandler {
-
+    @VisibleForTesting
+    void setHttpDummy(AsyncHttpClient httpClient) {
+        http = httpClient;
     }
 
     SourcePointClient(
@@ -82,11 +83,6 @@ class SourcePointClient {
         this.property = property;
         this.targetingParams = targetingParams;
         this.authId = authId;
-    }
-
-    @VisibleForTesting
-    void setHttpDummy(AsyncHttpClient httpClient) {
-        http = httpClient;
     }
 
     void getMessage(boolean isNative, String consentUUID, String meta, String euconsent, GDPRConsentLib.OnLoadComplete onLoadComplete) throws ConsentLibException {